optimization - using floats instead of doubles; more optimizations after profiling code;
git-svn-id: https://localhost/igvc@621 b8439097-6f30-0410-869d-b57f63bf1bb9

+++ 2009/candiii/Robot.h
@@ -104,7 +104,7 @@ public:
 
 	// avg heading control
 	// k_motors = the % of new value to use
-	double k_motors;
+	float k_motors;
 
 	// xml conf
 	void LoadXMLSettings();
